# [libc++][format] Switches to Unicode 15.1.